requirements.txt
streamlit
langchain-google-genai
python-dotenv
asyncio
browser-use
pydantic
